refactor: move resource and nullifier-key types into anoma-rm-core (wave 2)
Resource, ConsumedResourceWitness, the resource publics, NullifierKey,
NullifierKeyCommitment, and the LogicCircuit trait move to the core
crate; their sha256 derivations (commitment, nullifier, tag, nonce
derivation) come along, sourced from risc0-zkp. The engine-dependent
operations stay in anoma-rm-risc0 as free functions in the same-named
modules: resource::{create, reset_randomness, quantity_scalar, kind}
(rand + k256) and nullifier_key::random_pair (rand). Five call sites
migrate.
